Get Facebook icon by name, not resource ID
Although it is preferred to use the resource ID when loading resources, this makes it such that the Facebook SDK won't build out of the box, as it lacks an import statement to the project's "R" class. This fix looks up the Facebook icon by name, meaning that in order to integrate the Facebook SDK, one only needs to copy the icons to their project's "drawable" directories.