Allow dialogs to resize when the soft keyboard is shown.
FbDialog is a floating window containing a LinearLayout that contains a WebView. Since the window is floating, its dimensions are set to WRAP_CONTENT by default. Previously the LinearLayout's size was being fixed to the desired overall size of the dialog, and the WebView was set to FILL_PARENT. This worked fine, but would not allow the dialog window to resize when the soft keyboard is shown.

So, this change makes the LinearLayout's dimensions WRAP_CONTENT, fixes the width of the WebView to the desired width, and (after a little monkeying that I'll describe in the next paragraph), sets the height of the WebView to WRAP_CONTENT. This allows us to set the window's soft input mode to SOFT_INPUT_ADJUST_RESIZE so that when the soft keyboard is shown, the window shrinks to still fit in the visible space, reducing the height of the LinearLayout and the WebView, and allowing the WebView to scroll to reveal the content that before this code change would have been inaccessible because it would have been occluded by the soft keyboard.

As stated, this didn't quite work because before the WebView loads any content, its height is zero, making the window tiny (the height of the header) and look wrong. So, we initially fix the height of the WebView to our overall desired height, and then as soon as content is loaded, we change its height to WRAP_CONTENT.