Fix exception when trying to decode invalid string
This exception is barely seen in regular development, but if there is some problem with your corresponding Facebook App then Facebook will return URL fragments with a reference like "_", which caused this method to throw an uncaught ArrayIndexOutOfBoundsException. Since this method is splitting to key/values anyways, this little safety check will prevent the app from crashing.