Adding some try/catch blocks around these calls
If the user hits the "back" button on their device at the right time (or rather, the wrong time), then this can cause the FBDialog class to throw some strange exceptions about not being attached to the current activity, etc. Better to just put in a bit of extra protection here and swallow any exceptions that may result.