Remove the title for dialogs and add the 'x' on the top left corner
Summary: For the oauth, feed, request dialogs, removed the title bar and added
the 'x' button on the top left corner which when pressed dismisses the dialog.

Test Plan: Launched Hackbook and click on Login or requests, feed dialogs and
confirm that there is no title bar and the cross button is rightly displayed and
dismisses the dialog.

Reviewers: lshepard, yariv, jimbru, dkim, ttung, omids, raghuc1

Reviewed By: ttung

CC: caabernathy, brent, vksgupta, mattwkelly, lshepard, ttung

Differential Revision: 333017

Task ID: 700698