Fix for force closes where user rotates phone immediately after submitting status update.
There is a common case where a user rotates the phone immediately
after submitting a status update. This causes the dialog and spinner
to be detached from the window manager. Any subsequent calls to
dismiss() on either the dialog or the spinner will cause the whole
app to force close.

This is easily resolved by implementing the onDetachedFromWindow
callback, and then stopping the dismiss calls.