[android-sdk] Adds basic single-request logic to Request/Response classes.
Summary:
This commit changes our HTTP stack from the Apache HttpClient stack to the HttpURLConnection stack, which
the Android development team has suggested is the stack to use going forward. Basic functionality for single (i.e.,
non-batched) requests is functional, with many caveats. There is no (de-)serialization of GraphObjects as of yet, so most
requests to Graph API will not result in a Response that is useful for anything (though the debugger can be used to
verify that JSONObjects in the right shape are being created). There is no support for requests that return JSONArrays
rather than JSONObjects. There are a million "TODO port:" comments indicating where functional parity with iOS is still
lacking.

Test Plan:
- Ran unit tests

Revert Plan:

Reviewers: mmarucheck, jacl, gregschechte, ayden

Reviewed By: mmarucheck

CC: msdkexp@

Differential Revision: https://phabricator.fb.com/D501547