[android-sdk] Added support for collections in GraphObject wrapper.
Summary:
Added basic support for properties that get/set collections on a GraphObject.
Collections of primitive types and GraphObjects are supported. A GraphObjectCollection
class provides additional functionality beyond a Collection, including casting the
entire collection to a GraphObjectCollection of another GraphObject type. (Casting
collections of non-GraphObjects is not supported, as we don't proxy non-GraphObjects.)

TODOs remain around fully supporting putting collections, as we need to ensure that
we are storing the underlying JSONArray, if applicable.

Test Plan:
- Ran unit tests

Revert Plan:

Reviewers: jacl, mmarucheck, gregschechte, ayden

Reviewed By: mmarucheck

CC: msdkexp@, platform-diffs@lists

Differential Revision: https://phabricator.fb.com/D510578