[android-sdk] Added TestSession, FacebookTestCase, TestBlocker.
Summary:
This commit adds a TestSession class, which serves much the same purpose as the FBTestSession class in
the iOS SDK. It allows test accounts to be created in either a shared or private mode. In shared mode,
the permissions of the test account are encoded as part of its name, so future test cases can re-use
the same test account rather than create a new one. In private mode, a test account is created with the
necessary permissions and then deleted at the end of the session, to avoid proliferation of test accounts.
(The encoding of permissions is done using the String.hash method, which will be different than the encoding
used on iOS, so test accounts will not be shared between the two platforms at the moment, and there is a very
small possibility of hash code collisions that could result in permissions-related errors. The possibility is
pretty remote.)

Added a FacebookTestCase that simplifies use of TestSession by reading in applicationId and applicationSecret
from a config.json file that is created by the new configure_unit_tests.sh script, similar to what we did
for iOS.

TestBlocker provides message-pumping-and-waiting functionality similar to FBTestBlocker. Given the synchronous
nature of calls on Android, TestBlocker will not be needed as frequently as on iOS, but it is needed to test
session authorization, which happens asynchronously, as well as obviously any asynchronous-specific test cases.

Made necessary changes to some classes (AccessToken, Session) to make certain members have default visibility
rather than private, in order to let TestSession override functionality it needed to. Session can now take a
handler passed in from outside rather than using its own, which is necessary to correctly pump messages through
the TestBlocker's handler in test cases.

Fixed up some code in Response that was incorrectly handling Facebook returning naked 'true' or 'false' responses.

Test Plan:
- Ran all unit tests

Revert Plan:

Reviewers: mmarucheck, jacl, ayden, gregschechte

Reviewed By: mmarucheck

CC: msdkexp@, security-diffs@lists

Differential Revision: https://phabricator.fb.com/D514919