[android-sdk] Added more batch request tests, fixed some errors.
Summary:
Added more batch request unit tests, to bring us to parity with iOS.

Fixed an error in TestSession related to signed int handling.

Fixed an error in Response in dealing with results that are JSONObject.NULL.

Fixed an error in Request where we were serializing access_token parameters twice. Reordered serialization
of common parameters to be the same as in iOS (to simplify comparing URLs generated by the two different SDKs).

Added some helpers to FacebookTestCase and refactored some of the unit test code to increase sharing.

Test Plan:
- Ran all the tests

Revert Plan:

Reviewers: mmarucheck, jacl, gregschechte, ayden

Reviewed By: mmarucheck

CC: msdkexp@

Differential Revision: https://phabricator.fb.com/D531647