[android-sdk] Added support for REST requests and multiple results in Response.
Summary:
REST API support was another TODO for iOS parity. Added it.

Since REST APIs can return multiple results as a JSON array, rather than a JSON object with an
array-valued property, we need to support that. Added a getGraphObjectList accessor to Response.
A response can have either a GraphObject or a GraphObjectList, but not both.

Test Plan:
- Ran the tests

Revert Plan:

Reviewers: mmarucheck, jacl, gregschechte, ayden

Reviewed By: mmarucheck

CC: msdkexp@, security-diffs@lists

Differential Revision: https://phabricator.fb.com/D531849