Add support for Session.reauthorize()
Summary:
This adds reauthorize support to Session to enable changing the set
of permissions dynamically.

This also adds several unit tests for Session.  These tests caught
an issue in the SdkRuntime test that was leaving a modified static
that escaped the scope of the test, so that is fixed too.

Also, an unrelated cleanup to configure_unit_tests.sh to change
iOS->Android and fix indentation.

Test Plan:
- Unit tests

Revert Plan:

Tags:

Reviewers: clang, jacl

Reviewed By: clang

CC: msdkexp@

Differential Revision: https://phabricator.fb.com/D542934

Task ID: 1168401