[android-sdk] Fixes token extension
Summary:
There were a couple issues with token extension in a previous commit. A hard-coded 'true' should have been put back to
'false' before committing. The token refresh service name got messed up before committing. And Session should check to make
sure that it is not already attempting a token refresh (since it happens asynchronously) before starting another attempt.
Fixed the test infrastructure around this to be a little less hacky and provide a better test of the actual Session behavior.

Made a minor fix to Request to ensure that a request can't have both a REST method and a graph path.

Test Plan:
- Ran unit tests

Revert Plan:

Reviewers: mmarucheck, jacl, gregschechte, ayden

Reviewed By: mmarucheck

CC: msdkexp@

Differential Revision: https://phabricator.fb.com/D545627