[android-sdk] Improve performance of GraphObjectWrapper
Summary:
GraphObjectWrapper was using Method.equals to determine whether the invoked method was one it could handle. This
suffers from poor performance, and for our purposes it suffices to test whether the method names are equal (none of
the methods we implement are overloaded). This change improves performance of the FriendPickerView (in progress) by
approximately 300% when sorting 750 friends.

Test Plan:
- Ran unit tests

Revert Plan:

Reviewers: mmarucheck, jacl, gregschechte, ayden

Reviewed By: mmarucheck

CC: msdkexp@, platform-diffs@lists

Differential Revision: https://phabricator.fb.com/D550038