Create Login Button and SessionTracker
Summary: Creating a Login Button view and a SessionTracker to track either the active Session or a client-specified Session.

Test Plan: Tested with a modified version of the SessionLoginSample to make sure both the active session and client-specified sessions can be used. Also added unit tests for SessionTracker.

Reviewers: mmarucheck

Reviewed By: mmarucheck

CC: jacl

Differential Revision: https://phabricator.fb.com/D555686

Task ID: 1392265