Add support for request caching
Summary:
We now support caching the results from a batch of requests.

This adds a public RequestBatch class that implements List<Request>,
but also provides a scope for adding more metadata about the request.
This includes the Handler previously passed as a separate parameter,
as well as the new caching parameters.

RequestBatch.cacheKey controls the key used to load from and to save
to for cache results of a request.

RequestBatch.forceRoundTrip controls whether we do a roundtrip without
attempting to load from cache.  If cacheKey is set and forceRoundTrip
is true, we will fetch new results and update the cache.

A future change will reduce the number of overloads by requiring
RequestBatch for scenarios that involve batching or parameters related
to RequestBatch.

Test Plan:
Unit tests

Revert Plan:
ok to revert

Tags:

Reviewers: clang, mingfli, karthiks

Reviewed By: clang

CC: msdkexp@

Differential Revision: https://phabricator.fb.com/D554389

Task ID: 1406500