Added PictureDownloader
Summary:
This adds support for an internal DownloadManager class that allows
throttling and dynamically prioritizing Runnables.

GraphObjectAdapter and GraphObjectListFragment are modified to use
DownloadManager to throttle requests and to prioritize requests that
are relevant to items currently shown on the screen when scrolling
stops.

Test Plan:
- unit tests

Revert Plan:
- safe to revert

Reviewers: clang, karthiks, mingfli

Reviewed By: clang

CC: msdkexp@, platform-diffs@lists

Differential Revision: https://phabricator.fb.com/D559726

Task ID: 1424276