[android-sdk] Adds paging to place picker, filtering.
Summary:
Added a setFilter and associated methods to GraphObjectListFragment to allow consumers to specify filters that will be
applied to objects prior to their being displayed.

Implemented as-needed paging in place picker, so results are loaded only when needed (i.e., when scrolled to the bottom of
the list).

Fixed a bug in TestSession that was deleting test accounts when it shouldn't have.

Test Plan:
- Ran unit tests
- Ran FriendPickerSample
- Ran PlacePickerSample

Revert Plan:

Reviewers: mmarucheck, mingfli

Reviewed By: mmarucheck

CC: msdkexp@, platform-diffs@lists, zizhuang

Differential Revision: https://phabricator.fb.com/D564065