[android-sdk] Add caching support to the picker fragments.
Summary:
Friend and place picker will now use the cache to retrieve results, if possible. If cached results are used, a "refresh"
query is executing after a short timeout (currently 2 seconds) to ensure that changes on the service are reflected locally.

Fixed Response to not reference HttpURLConnection methods for cached results, as these would end up blocking until the
request finished anyway.

Added a workaround for an Android framework bug relating to AsyncTask in FacebookActivityTestCase.

Test Plan:
- Ran unit tests
- Ran FriendPickerSample
- Ran PlacePickerSample

Revert Plan:

Reviewers: mmarucheck, mingfli

Reviewed By: mmarucheck

CC: msdkexp@, platform-diffs@lists

Differential Revision: https://phabricator.fb.com/D566767