Preliminary work towards making Session serializable.
Summary:
We need to have Session serializable in order to support a
more androidey api, since in Android the application might be killed
while waiting for the login activity result.

There's probably a lot more work toward getting that final goal, but
this is a reasonable first step.

(I chose Serializable over Parcelable, because it's far more easier to
implement and it's not performacne critical for now.)

I also removed the final fields from the Session object because it's hard
to work with them in the context of objects being serializable. Also, it's
hard to test in the long run.

Test Plan:
Wrote a basic unit test to make sure that no crashes happen
during the serialization and unserialization processes.

Revert Plan: OK to revert.

Reviewers: mmarucheck

Reviewed By: mmarucheck

CC: jakubv, msimpson

Differential Revision: https://phabricator.fb.com/D567475

Task ID: 1387452