Fixing crash when user cancels the SSO signin activity
Summary:
The SSO authorization was validating that the result bundle from the
SSO activity is not null in all cases. However, when the user presses back
and cancels the signin, the bundle is null.

Test Plan:
Tested with a simple app that pressing the back button on the signin activity
returns the app back to the state befor the SSO was started

Reviewers: mmarucheck

Reviewed By: mmarucheck

CC: peterb, seansible-diffs@

Differential Revision: https://phabricator.fb.com/D569512

Blame Revision: D536878

Task ID: 1451653