Fix GraphObjectWrapper so that Iterables can be serialized to a request.
Summary: GraphObjectWrapper wasn't properly setting Iterable/Collection/List types, and as a result they weren't being serialized in the outgoing requests. This fixes the issue. Also changed the behavior of the "Log out" button in LoginFragment so that it shows a confirmation dialog.

Test Plan: Added a unit test, ran Scrumptious on emulator and verified with test users that they're tagged in newsfeed.

Reviewers: mmarucheck, clang, karthiks

Reviewed By: clang

CC: msdkexp@, platform-diffs@lists

Differential Revision: https://phabricator.fb.com/D585535

Task ID: 1765303