Deprecate overloading request code and login behavior
Summary:
The Facebook class's most full authorize() method takes a single
parameter that encodes both the request code to use for SSO and
login behavior.  There are two problems with this:
1) There is no way to specify SSO_ONLY
2) The semantics of the parameter is mixed

While it may not be possible to use -1 as a valid request code, (is
it?) it is better to have an API that does not convolute these two
concerns.

Although Facebook will be deprecated and we do not intend to add
features going forward, creating the Session via Facebook is the
mechanism by which apps can take a dependency on new functionality
incrementally.  To support this, we are adding an authorize()
overload here.

Test Plan:
- run Hackbook

Revert Plan:
- safe

Tags:

Reviewers: francip, clang, mingfli

Reviewed By: francip

CC: msdkexp@

Differential Revision: https://phabricator.fb.com/D585977

Task ID: 1430683