[android-sdk] Fix null reference in pickers.
Summary:
If loadData was called on a picker in a detached state, we would attempt to dereference a null loadingStrategy.loader.
Fixed to not do that.

Also fixed a couple issues with Scrumptious sample where we called loadData unnecessarily and left LocationListeners
hanging around when they were no longer useful.

Test Plan:
- Ran Scrumptious

Revert Plan:

Reviewers: mmarucheck, mingfli, karthiks

Reviewed By: mmarucheck

CC: platform-diffs@lists

Differential Revision: https://phabricator.fb.com/D594417

Task ID: 1772164, 1792971, 1793005, 1793090