Fixing profile picture view for pending requests
Summary:
Profile picture view was not saving necessary state when there was a pending request. This resulted in the restore instance not refreshing the image. This fix addresses that issue.

Additionally, also making sure that response handlers are not executing for orphaned ProfilePictureViews

Also, updating the sample to have Chris' id.

Test Plan: Tested Sample

Reviewers: mmarucheck, mingfli, clang

Reviewed By: mmarucheck

CC: msdkexp@

Differential Revision: https://phabricator.fb.com/D594872

Task ID: 1793014