[android-sdk] Fix PlacePickerFragment search text timer to avoid null reference
Summary:
The timer handler assumed getActivity() would return non-null. Because the handler is running
on a background thread, it is possible for there to be no activity if, e.g., the device is being
rotated at the time. Fixed the handler to just do what Activity.runOnUiThread() does, removing the
reliance on getActivity() returning non-null.

Test Plan:
- Ran PlacePickerSample, did lots of edits in the search box while rotating the device

Revert Plan:

Reviewers: mmarucheck, mingfli, karthiks

Reviewed By: mmarucheck

Differential Revision: https://phabricator.fb.com/D595143

Task ID: 1793786