Fix Scrumptious crash during SSO reauthorize
Summary: Fixes 3 bugs - 1. Session doesn't serialize the pending request; 2. if reauthorize fails, we currently put the Session in CLOSED_LOGIN_FAILED state, which is inaccurate; 3. also in the onActivityResult in Scrumptious, we're only calling session.onActivityResult for the success case when we should be calling it in failures as well.

Test Plan: tested on emulator

Reviewers: karthiks, chrisp, mmarucheck

Reviewed By: karthiks

CC: msdkexp@

Differential Revision: https://phabricator.fb.com/D597497

Task ID: 1799613