Fix exception when rotating device during Dialog auth.
Summary: Several scenarios can cause the dialog's containing Activity to be destroyed, causing a window leak. This fix starts another activity that handles the dismissal of dialogs if the activity is about to go away. Note that developers must add this Activity to their manifest, so this still falls back to the old behavior if they forget to add it.

Test Plan: tested on emulator with both SSO and dialog auth.

Reviewers: mmarucheck, karthiks, clang

Reviewed By: mmarucheck

CC: msdkexp@, caabernathy

Differential Revision: https://phabricator.fb.com/D598624

Task ID: 1768834