[android-sdk] Fix pickers to not display title bar after re-creation.
Summary:
GraphObjectListFragment was inflating its title bar (if appropriate) in onCreateView. But if the fragment
is being re-created, we have not yet de-serialized our settings form the bundle yet, so the default (true) was being
used. Moved this to onActivityCreated, at which point we know whether or not to display it.

Test Plan:
- Ran BooleanOG (no title bar) and FriendPickerSample (title bar) on tablet, tried rotating tablet, observed
correct behavior in both.

Revert Plan:

Reviewers: mmarucheck, mingfli, karthiks

Reviewed By: mingfli

CC: platform-diffs@lists

Differential Revision: https://phabricator.fb.com/D603502

Task ID: 1816646