[android-sdk] Fix HelloFacebookSample to be more robust to ordering of callbacks
Summary:
HelloFacebookSample used the idea of a "pending action" to perform actions such as posting status updates or photos
after, e.g., a permissions upgrade. But posting the status update also relies on having a valid 'user' instance
variable set. Following a permissions upgrade, this value may be null because the query to retrieve 'me' has not
yet finished. Adding an additional check for this case, and checking for any pending action after we retrieve 'me'.

Test Plan:
- Ran HelloFacebook (without SSO), was able to upgrade permissions and status update was posted successfully
- Un-TOSed app and re-ran, app did not crash (this is still not completely desirable behavior, as we should check
for the un-TOS case and do something else, but it fixes the crash)

Revert Plan:

Reviewers: mmarucheck, mingfli, karthiks

Reviewed By: mmarucheck

Differential Revision: https://phabricator.fb.com/D603781

Task ID: 1816893, 1816912