[android-sdk] Fix place picker so search box is created after title bar.
Summary:
Previous commit changed when we create the title bar in the pickers; place picker depends on knowing whether or
not there is a title bar to determine whether it needs to layout the search box below it or not. With the other
change, we were prematurely deciding there was no title bar, so the search box was at the top of the parent
view, overlapping with the title bar. Changed it to the be created at the same step of the lifecycle as the title bar.

Test Plan:
- Ran Scrumptious & PlacePickerSample, observed search box in the correct spot, including after rotations

Revert Plan:

Reviewers: mmarucheck, mingfli, karthiks

Reviewed By: mmarucheck

Differential Revision: https://phabricator.fb.com/D603860

Blame Revision: D603502

Task ID: 1817510