[android-sdk] Added Request.newUploadVideoRequest helper to set up for video upload.
Summary:
newUploadVideoRequest returns a Request that has been initialized to upload a video; the video is
specified via a File object passed to the method. The name of the underlying file is used as the name
of the uploaded attachment, so its extension should be appropriate to the format of the video file.

Test Plan:
- Added new unit test to upload a short video of a mysterious window in Palo Alto, CA.
- Verified after running unit test that video was encoded and displayed in test user's activity log.

Revert Plan:

Reviewers: mmarucheck, mingfli, karthiks

Reviewed By: mmarucheck

Differential Revision: https://phabricator.fb.com/D607644

Task ID: 1441128