[android-sdk] Fixes several issues with picker samples and authentication.
Summary:
Both FriendPickerSample and PlacePickerSample tried to open a new active session every time their activity was
created. In non-SSO cases, this caused the login and TOS pages to be displayed twice. With the new logic, we will
try to re-open the session only if it is closed, so un-TOSing the app and re-running it will display an error the first
time through saying that the app is not authorized, then try to re-authenticate the user when they return to the main
activity.

FriendPickerSample now re-draws the list of selected friends when the activity is re-created, e.g., after rotation.

Test Plan:
- Ran both samples (in non-SSO mode), tried un-TOSing and re-running app, rotating app, etc.

Revert Plan:

Reviewers: mmarucheck, mingfli, karthiks

Reviewed By: mmarucheck

Differential Revision: https://phabricator.fb.com/D607705

Task ID: 1816798, 1816774, 1817045