Use executeOnExecutor when available                                                                           [0/44111] Summary: We want to allow Requests to run in parallel on multiple threads. However, this is a summary of behavior at different API-lev
Summary:
To enable this, we attempt to use executeOnExecutor if it is available,
and call it with AsyncTask.THREAD_POOL_EXECUTOR (or an Executor the app
specifies via Settings.setExecutor).

If that API is not available, we fall back to AsyncTask.execute.

Either way, for API-levels >=8 (which is what we support), we are
running Requests in parallel by default, and we provide a mechanism
for overriding this behavior whenever possible.

Test Plan:
- Unit tests
- Set a breakpoint for a test that exercises this and watch behavior
  for android-8 and android-16.

Revert Plan:
- safe

Tags:

Reviewers: clang, mingfli

Reviewed By: clang

CC: msdkexp@

Differential Revision: https://phabricator.fb.com/D607955