[android-sdk] Fix authentication logic in JustRequestSample.
Summary:
JustRequestSample contained logic that assumed the same Activity was present after authentication as before. Since
it is possible for the Activity to be recycled, it needs to be resilient to this case. Changed auth logic to not
assume that same Activity and remember whether a request was pending following authentication.

Test Plan:
- Ran JustRequestSample without Facebook app installed, verified that after logging in and TOSing app, pending request
was executed and results displayed.

Revert Plan:

Reviewers: mmarucheck, mingfli, karthiks

Reviewed By: mmarucheck

Differential Revision: https://phabricator.fb.com/D608734

Task ID: 1816970