[android-sdk] Added WebDialog class, deprecated FbDialog, which now is a thin wrapper around WebDialog.
Summary:
As we are deprecating classes and methods in the com.facebook.android namespace, including FbDialog, we add a
new class, WebDialog, in the com.facebook namespace. It has functionality similar to FbDialog, but includes a number of
helper methods to construct specific types of Facebook Web dialogs and has a simpler listener interface.

FbDialog is now a thin wrapper around WebDialog that preserves existing callback semantics.

Test Plan:
- Ran Hackbook, verified that dialogs display successfully (using both FbDialog and WebDialog)
- Ran several other samples, verified that Web-based authentication continues to work

Revert Plan:

Reviewers: mmarucheck, mingfli, karthiks

Reviewed By: mmarucheck

CC: ekoneil, caabernathy

Differential Revision: https://phabricator.fb.com/D610413

Task ID: 1831402