[android-sdk] Add profile picture pre-fetching to GraphObjectAdapter for smoother scrolling.
Summary:
To provide a smoother scrolling experience, GraphObjectAdapter will now prefetch a few profile pictures on either
side of the visible items, in the hope that by the time those items scroll into view, the picture will be in memory and
can be immediately displayed rather than showing a default picture for a split second before the profile picture appears.
Image cache misses or large jumps in the position of the list view will still result in the default picture being
displayed.

Test Plan:
- Ran FriendPickerSample with a large number of friends, scrolled up and down and did not observe default picture
being displayed when in a section of friends with cached bitmaps.

Revert Plan:

Reviewers: mmarucheck, karthiks, mingfli

Reviewed By: karthiks

CC: platform-diffs@lists

Differential Revision: https://phabricator.fb.com/D612678

Task ID: 1788580