[android-sdk] Added SessionLoginFragmentSample, renamed SessionLoginSample to SessionLoginActivitySample
Summary:
SessionLoginFragmentSample looks and acts identically to SessionLoginActivitySample (formerly just SessionLoginSample), but
is implemented in terms of a Fragment within a FragmentActivity rather than by a simple Activity. It demonstrates using
the overloaded methods in Session that take a Fragment rather than an Activity, and how to pass data from onActivityResult
to Session to handle authorization.

Test Plan:
- Ran both samples, verified behavior was the same in both of them

Revert Plan:

Reviewers: mmarucheck, mingfli, karthiks

Reviewed By: mmarucheck

CC: caabernathy, ekoneil, subodh

Differential Revision: https://phabricator.fb.com/D613975

Task ID: 1833667