[android-sdk] Make HelloFacebook more robust when handling pending actions.
Summary:
Previously, a race condition existed in HelloFacebook where handlePendingActions could be called before all conditions
necessary to successfully call the pending action were in place. In particular, a cancelled permissions upgrade would
cause us to attempt to post a photo or status update anyway. Added a handler so that cancelling the permissions upgrade
will display an informative message and abort the attempt to post. Improved the pending action handlers so they will not
attempt an action if all necessary conditions are not met, but will instead stay in the pending state until they are
called again. (For instance, we may be called because this.user was filled in, while we are still waiting for a pending
permissions upgrade -- now, we'll recognize that we aren't ready to execute the pending action and will wait for the
permissions upgrade to complete.)

Test Plan:
- Ran HelloFacebook on device with Facebook app installed and "Don't keep activities" developer option on
- Logged on as user who had not TOSed app
- Clicked "Post Photo"
- Hit back button rather than agreeing to permissions upgrade
- Saw warning message
- Clicked "Post Photo" again
- Accepted permissions upgrade
- Saw that photo was posted upon returning to app

Revert Plan:

Reviewers: mmarucheck, mingfli, karthiks, gregschechte

Reviewed By: mingfli

Differential Revision: https://phabricator.fb.com/D614469

Task ID: 1846373