[android-sdk] Removes PickerFragment, makes GraphObjectListFragment public, rearchitect HelloFacebookSample
Summary:
The rationale for having a separate PickerFragment interface that GraphObjectListFragment extended is gone, so there is
no reason not to just make GraphObjectListFragment public, which is not only simpler, but allows use of the picker fragments
as-a Fragment in a polymorphic way, which may be useful for certain apps.

Modified fragment listeners to take the fragment as the first parameter, to simplify implementation of callbacks.

HelloFacebookSample now has a single Activity which swaps in and out the picker fragments, rather than wrapping them in
separate activities. Also cleaned up a couple other issues with HelloFacebookSample.

Test Plan:
- Ran unit tests
- Ran all samples that use pickers

Revert Plan:

Reviewers: mmarucheck, mingfli, karthiks, gregschechte

Reviewed By: mmarucheck

CC: ekoneil, caabernathy, platform-diffs@lists

Differential Revision: https://phabricator.fb.com/D616779

Task ID: 1852204, 1846439, 1846399