[android-sdk] Add better error handling to WebDialog.
Summary:
WebDialog parsed out OAuth-related errors, but did not parse other types of errors that might be returned from a
dialog. Now we parse them and create a FacebookRequestError populated with as much information as we got from the
dialog.

Test Plan:
- Modified Hackbook to call dialogs with parameters that would cause errors, verified that error object was
populated with the information we got back.

Revert Plan:

Reviewers: mmarucheck, mingfli, karthiks

Reviewed By: mmarucheck

Differential Revision: https://phabricator.fb.com/D618163

Task ID: 1768823