[android-sdk] Forcibly hide virtual keyboard when detaching PlacePickerFragment.
Summary:
At least on some API levels, simply detaching the PlacePickerFragment is not enough to cause the system to
dismiss the virtual keyboard if it was being displayed because the user had focus in the search box. Overriding
onAttach/onDetach to explicitly set the soft input mode for the search box.

Test Plan:
- Ran HelloFacebook on 2.3.3 emulator
- Brought up place picker
- Clicked in search box, waited for keyboard to appear
- Clicked Done, and keyboard disappeared, where previously it would remain on-screen after the picker disappeared

Revert Plan:

Reviewers: mmarucheck, mingfli, karthiks, gregschechte

Reviewed By: mmarucheck

Differential Revision: https://phabricator.fb.com/D618141

Task ID: 1853703